[FIX] Og_Subscription/js/switch-subscription deps
Fix Ordergroove_Subscription/js/switch-subscription module definition dependencies requiring `window.checkoutConfig` causing errors.

Fix [Issue 24](https://github.com/ordergroove/magento-module-24/issues/24).

Only load models that require `window.checkoutConfig` if it is available.
Stop errors caused by `window.checkoutConfig` being undefined in Magento_Checkout/js/model/quote & Magento_Checkout/js/model/cart/totals-processor/default.

The workflow that uses these models is the same the click event listeners for both the 'og-optin-button' & 'og-optout-button', with the expectation that the page URL path begins with "checkout".
Add `estimateTotals` function that runs the cart totals-processor `estimateTotals()` workflow, but only if the `window.checkoutConfig` is available, as it would be within the already checked for "checkout" URL path pages.  Only require classes that depend on `window.checkoutConfig` if it is defined.  Do nothing  if it is not defined.
This does not change any of the current expected functionality of the switch-scripts file, continuing to run these `estimateTotals()` calls where they would run before.
It does prevent attempting to load models that are unavailable due to missing required data dependencies, preventing calls that cannot be made anyway and the errors that accompany those bad calls.  The data is available and the calls are made in the places where it was expected to be available before.